chore(ZN-386): update Cypress to 10.11
