+++ .github/workflows/e2e.yml
@@ -5,7 +5,7 @@ jobs:
   prepare_app:
     runs-on: ubuntu-latest
     container:
-      image: cypress/included:8.7.0
+      image: cypress/included:10.11.0
       options: --user 1001
     steps:
       - uses: actions/checkout@master
