--- e2e/cypress.json
@@ -1,3 +1,0 @@
-{
-
-}
