+++ e2e/package.json
@@ -18,7 +18,7 @@
   "dependencies": {
     "@cypress/webpack-preprocessor": "^5.9.1",
     "chokidar": "^3.5.2",
-    "cypress": "8.7.0",
+    "cypress": "10.11.0",
     "dotenv-safe": "^8.2.0",
     "ts-loader": "^9.2.6",
     "typescript": "^4.4.4",
