+++ e2e/scripts/watcher/watcher.config.ts
@@ -2,8 +2,8 @@ import { WatcherConfig } from '../scripts.types';
 
 const config: WatcherConfig = {
   pathFromGitToCurrentPackageJson: 'e2e/',
-  testsDir: `./cypress/integration`,
-  testFilesPattern: 'e2e\\/cypress\\/integration\\/.*?(?=.test).*?.ts',
+  testsDir: `./cypress/e2e`,
+  testFilesPattern: 'e2e\\/cypress\\/e2e\\/.*?(?=.cy).*?.ts',
   browser: 'chrome',
 };
 export default config;
